requirements.txt
cattrs
colorama
pyyaml
